📦 Remove example, icons and type main files
