--- src/examples/App.scss
@@ -1,70 +1,0 @@
-.title {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-main {
-  font-family: Roboto, "Helvetica Neue", sans-serif;
-  box-sizing: border-box;
-  height: 100vh;
-  padding: 40px;
-  background-color: #8E95B0;
-  
-  > .container {
-	display: flex;
-	justify-content: space-around;
-	margin-top: 40px;
-  }
-}
-
-.section {
-  display: flex;
-  justify-content: center;
-  width: 400px;
-  height: 400px;
-  border-radius: 8px;
-  
-  &.basic {
-	background-color: #DAD2D0;
-  }
-  
-  &.rhf {
-	color: #FFFEFC;
-	background-color: #071328;
-	
-	& > form > input:last-child {
-	  font-weight: bold;
-	  background-color: #EC5990;
-	}
-  }
-  
-  > form {
-	display: flex;
-	flex-direction: column;
-	width: 80%;
-	height: fit-content;
-	padding: 16px;
-	
-	> label {
-	  margin-top: 24px;
-	  margin-bottom: 8px;
-	}
-	
-	> input {
-	  height: 24px;
-	  border: none;
-	  border-radius: 6px;
-	}
-	
-	& > input:last-child {
-	  width: 72px;
-	  height: 32px;
-	  margin-top: 40px;
-	  margin-bottom: 16px;
-	  padding: 8px 12px;
-	  color: #FFFEFC;
-	  background-color: #262627;
-	}
-  }
-}
