--- src/lib/components/variables.scss
@@ -1,21 +1,0 @@
-
-[theme='light'] {
-  --BG-COLOR: #FFFEFC;
-  --FONT-COLOR-strong: #262627;
-  --FONT-COLOR-light: #616060;
-  --FONT-COLOR-otherMonth: #616161;
-  --FONT-COLOR-disabled: #9E9E9E;
-  --SELECTED-COLOR-default: #FFFEFC;
-  --SELECTED-COLOR-hover: #CCC;
-  
-}
-
-[theme='dark'] {
-  --SELECTED-COLOR-hover: #3A3A3A;
-  --SELECTED-COLOR-default: rgb(38, 38, 39);
-  --BG-COLOR: #262627;
-  --FONT-COLOR-strong: #FFFEFC;
-  --FONT-COLOR-otherMonth: #6A6A6A;
-  --FONT-COLOR-light: #BCBCBC;
-  --FONT-COLOR-disabled: #373737;
-}
