--- src/lib/index.js
@@ -1,3 +1,0 @@
-import Datepicker from './components/Datepicker'
-
-export default Datepicker
